Report failure even when CBMC crashes, refactor CBMC parser/renderer
