+++ kani-driver/src/call_cbmc.rs
@@ -4,36 +4,59 @@
 use anyhow::{bail, Result};
 use kani_metadata::HarnessMetadata;
 use std::ffi::OsString;
+use std::fmt::Write;
 use std::path::Path;
 use std::process::Command;
-use std::time::Instant;
+use std::time::{Duration, Instant};
 
-use crate::args::KaniArgs;
-use crate::cbmc_output_parser::{process_cbmc_output, VerificationOutput};
+use crate::args::{KaniArgs, OutputFormat};
+use crate::cbmc_output_parser::{
+    process_cbmc_output, CheckStatus, ParserItem, Property, VerificationOutput,
+};
+use crate::cbmc_property_renderer::{format_result, kani_property_filter};
 use crate::session::KaniSession;
 
-#[derive(PartialEq, Eq)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum VerificationStatus {
     Success,
     Failure,
 }
 
+/// Our (kani-driver) notions of CBMC results.
+///
+/// `messages` will have `Result` removed, and that can instead be found directly in `results`.
+#[derive(Debug)]
+pub struct VerificationResult {
+    /// Whether verification should be considered to have succeeded, or have failed.
+    pub status: VerificationStatus,
+    /// The parsed output, message by message, of CBMC. `Result` has been removed, however.
+    pub messages: Option<Vec<ParserItem>>,
+    /// The `Result` properties in detail.
+    pub results: Option<Vec<Property>>,
+    /// CBMC process exit status. NOTE: Only potentially useful if `status` is `Failure`.
+    /// Kani will see CBMC report "failure" that's actually success (interpreting "failed"
+    /// checks like coverage as expected and desirable.)
+    pub exit_status: i32,
+    /// The runtime duration of this CBMC invocation.
+    pub runtime: Duration,
+}
+
 impl KaniSession {
     /// Verify a goto binary that's been prepared with goto-instrument
-    pub fn run_cbmc(&self, file: &Path, harness: &HarnessMetadata) -> Result<VerificationStatus> {
+    pub fn run_cbmc(&self, file: &Path, harness: &HarnessMetadata) -> Result<VerificationResult> {
         let args: Vec<OsString> = self.cbmc_flags(file, harness)?;
 
         // TODO get cbmc path from self
         let mut cmd = Command::new("cbmc");
         cmd.args(args);
 
-        let now = Instant::now();
+        let start_time = Instant::now();
 
-        let verification_output = if self.args.output_format == crate::args::OutputFormat::Old {
+        let verification_results = if self.args.output_format == crate::args::OutputFormat::Old {
             if self.run_terminal(cmd).is_err() {
-                VerificationOutput { status: VerificationStatus::Failure, processed_items: None }
+                VerificationResult::mock_failure()
             } else {
-                VerificationOutput { status: VerificationStatus::Success, processed_items: None }
+                VerificationResult::mock_success()
             }
         } else {
             // Add extra argument to receive the output in JSON format.
@@ -43,25 +66,22 @@ impl KaniSession {
             // Spawn the CBMC process and process its output below
             let cbmc_process_opt = self.run_piped(cmd)?;
             if let Some(cbmc_process) = cbmc_process_opt {
-                // The introduction of reachability checks forces us to decide
-                // the verification result based on the postprocessing of CBMC results.
-                process_cbmc_output(
-                    cbmc_process,
-                    self.args.extra_pointer_checks,
-                    &self.args.output_format,
-                )
+                let output = process_cbmc_output(cbmc_process, |i| {
+                    kani_property_filter(
+                        i,
+                        self.args.extra_pointer_checks,
+                        &self.args.output_format,
+                    )
+                })?;
+
+                VerificationResult::from(output, start_time)
             } else {
-                VerificationOutput { status: VerificationStatus::Failure, processed_items: None }
+                VerificationResult::mock_failure()
             }
         };
-        // TODO: We should print this even the verification fails but not if it crashes.
-        if !self.args.dry_run {
-            let elapsed = now.elapsed().as_secs_f32();
-            println!("Verification Time: {}s", elapsed);
-        }
 
-        self.gen_and_add_concrete_playback(harness, &verification_output)?;
-        Ok(verification_output.status)
+        self.gen_and_add_concrete_playback(harness, &verification_results)?;
+        Ok(verification_results)
     }
 
     /// used by call_cbmc_viewer, invokes different variants of CBMC.
@@ -161,6 +181,94 @@ impl KaniSession {
         }
 
         args
+    }
+}
+
+impl VerificationResult {
+    /// Computes a `VerificationResult` (kani-driver's notion of the result of a CBMC call) from a
+    /// `VerificationOutput` (cbmc_output_parser's idea of CBMC results).
+    ///
+    /// NOTE: We actually ignore the CBMC exit status, in favor of two checks:
+    ///   1. Examining the actual results of CBMC properties.
+    ///       (CBMC will regularly report "failure" but that's just our cover checks.)
+    ///   2. Positively checking for the presence of results.
+    ///       (Do not mistake lack of results for success: report it as failure.)
+    fn from(output: VerificationOutput, start_time: Instant) -> VerificationResult {
+        if let Some(mut items) = output.processed_items {
+            let result_idx = items.iter().position(|x| matches!(x, ParserItem::Result { .. }));
+            if let Some(result_idx) = result_idx {
+                let result = items.remove(result_idx);
+                if let ParserItem::Result { result } = result {
+                    let number_failed_properties =
+                        result.iter().filter(|prop| prop.status == CheckStatus::Failure).count();
+                    let status = if number_failed_properties == 0 {
+                        VerificationStatus::Success
+                    } else {
+                        VerificationStatus::Failure
+                    };
+                    VerificationResult {
+                        status,
+                        messages: Some(items),
+                        results: Some(result),
+                        exit_status: output.process_status,
+                        runtime: start_time.elapsed(),
+                    }
+                } else {
+                    unreachable!() // We filtered for this to be true
+                }
+            } else {
+                // We never got results from CBMC - also shouldn't really happen unless something went wrong
+                VerificationResult {
+                    status: VerificationStatus::Failure,
+                    messages: Some(items),
+                    results: None,
+                    exit_status: output.process_status,
+                    runtime: start_time.elapsed(),
+                }
+            }
+        } else {
+            // We don't have ANY messages to report - shouldn't really happen unless something went wrong
+            VerificationResult {
+                status: VerificationStatus::Failure,
+                messages: None,
+                results: None,
+                exit_status: output.process_status,
+                runtime: start_time.elapsed(),
+            }
+        }
+    }
+    pub fn mock_success() -> VerificationResult {
+        VerificationResult {
+            status: VerificationStatus::Success,
+            messages: None,
+            results: None,
+            exit_status: 42,
+            runtime: Duration::from_secs(0),
+        }
+    }
+    fn mock_failure() -> VerificationResult {
+        VerificationResult {
+            status: VerificationStatus::Failure,
+            messages: None,
+            results: None,
+            exit_status: 42,
+            runtime: Duration::from_secs(0),
+        }
+    }
+
+    pub fn render(&self, output_format: &OutputFormat) -> String {
+        if let Some(results) = &self.results {
+            let show_checks = matches!(output_format, OutputFormat::Regular);
+            let mut result = format_result(results, show_checks);
+            writeln!(result, "Verification Time: {}s", self.runtime.as_secs_f32()).unwrap();
+            result
+        } else {
+            let verification_result = console::style("FAILED").red();
+            format!(
+                "\nCBMC failed with status {}\nVERIFICATION:- {}\n",
+                self.exit_status, verification_result
+            )
+        }
     }
 }
